Refactor code even more: - Separate FileRepositoryImpl from FileRepository interface - Move SortOrder to models folder - Move Generators to models folder
