+++ app/src/main/kotlin/de/ywegel/svenska/data/VocabularyRepository.kt
@@ -2,6 +2,7 @@
 
 package de.ywegel.svenska.data
 
+import de.ywegel.svenska.data.model.SortOrder
 import de.ywegel.svenska.data.model.Vocabulary
 import kotlinx.coroutines.flow.Flow
