+++ app/src/main/kotlin/de/ywegel/svenska/di/SvenskaModule.kt
@@ -18,7 +18,6 @@ import dagger.hilt.components.SingletonComponent
 import de.ywegel.svenska.data.ContainerRepository
 import de.ywegel.svenska.data.FavoritesAndPronunciationsRepository
 import de.ywegel.svenska.data.FileRepository
-import de.ywegel.svenska.data.FileRepositoryImpl
 import de.ywegel.svenska.data.SearchRepository
 import de.ywegel.svenska.data.VocabularyRepository
 import de.ywegel.svenska.data.db.ContainerDao
@@ -27,6 +26,7 @@ import de.ywegel.svenska.data.db.VocabularyDao
 import de.ywegel.svenska.data.db.VocabularyDatabase
 import de.ywegel.svenska.data.impl.ContainerRepositoryImpl
 import de.ywegel.svenska.data.impl.FavoritesAndPronunciationsRepositoryImpl
+import de.ywegel.svenska.data.impl.FileRepositoryImpl
 import de.ywegel.svenska.data.impl.SearchRepositoryImpl
 import de.ywegel.svenska.data.impl.VocabularyRepositoryImpl
 import de.ywegel.svenska.data.preferences.OVERVIEW_PREFERENCES_NAME
