+++ app/src/main/kotlin/de/ywegel/svenska/ui/common/ClickableText.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import de.ywegel.svenska.data.GeneratorConstants
+import de.ywegel.svenska.data.model.GeneratorConstants
 import de.ywegel.svenska.ui.theme.Spacings
 import de.ywegel.svenska.ui.theme.SvenskaTheme
