+++ app/src/main/kotlin/de/ywegel/svenska/ui/common/ContainerSelection.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import de.ywegel.svenska.data.containers
 import de.ywegel.svenska.data.model.VocabularyContainer
+import de.ywegel.svenska.data.model.containers
 import de.ywegel.svenska.ui.theme.Spacings
 import de.ywegel.svenska.ui.theme.SvenskaIcons
 import de.ywegel.svenska.ui.theme.SvenskaTheme
