+++ app/src/main/kotlin/de/ywegel/svenska/ui/overview/OverviewScreen.kt
@@ -48,7 +48,7 @@ import com.ramcosta.composedestinations.generated.destinations.WordGroupsScreenD
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import de.ywegel.svenska.R
 import de.ywegel.svenska.data.model.Vocabulary
-import de.ywegel.svenska.data.vocabularies
+import de.ywegel.svenska.data.model.vocabularies
 import de.ywegel.svenska.navigation.SvenskaGraph
 import de.ywegel.svenska.ui.common.HorizontalSpacerS
 import de.ywegel.svenska.ui.common.IconButton
